requirements.txt
azure.core
azure-ai-documentintelligence
streamlit
azure-storage-blob
azure-identity
python-dotenv

